requirements.txt
django~=4.1.7
django-debug-toolbar~=3.8.1

djangorestframework~=3.14.0
google-cloud-secret-manager
psycopg2
